move store macros into store
