+++ src/local_array/store/mod.rs
@@ -1,7 +1,10 @@
 pub mod custom_alloc;
 pub mod iterators;
+pub mod errors;
 
 pub(crate) mod default_store;
 pub(crate) mod atomic_types;
 
 pub use default_store::DefaultStore;
+#[macro_use]
+mod macros;
